ADD: logout and settings menu
