+++ Beakon/app/src/main/java/erica/beakon/Pages/MyMovementsTab.java
@@ -1,30 +1,20 @@
 package erica.beakon.Pages;
 
-import android.app.Fragment;
-import android.content.Intent;
-import android.graphics.Color;
+
 import android.os.Bundle;
-import android.support.v4.app.FragmentTransaction;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.ImageButton;
 import android.widget.ListView;
 import android.widget.TextView;
 
-import com.facebook.login.LoginManager;
 import com.google.firebase.database.ChildEventListener;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.ValueEventListener;
 
-import java.util.HashMap;
-
 import erica.beakon.Adapters.MyMovementAdapter;
-import erica.beakon.LoginPage;
-import erica.beakon.MainActivity;
 import erica.beakon.Objects.Movement;
 import erica.beakon.R;
 
@@ -32,11 +22,8 @@ public class MyMovementsTab extends MovementsTab {
 
     public static final String TAG = "MY MOVEMENTS TAB";
     MyMovementAdapter adapter;
-    View view;
     ListView listView;
     TextView message;
-    Button logoutButton;
-    Button userPrefButton;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
@@ -46,41 +33,14 @@ public class MyMovementsTab extends MovementsTab {
 
         listView = (ListView) view.findViewById(R.id.my_movements_list);
         message = (TextView) view.findViewById(R.id.no_movments_message);
-        logoutButton = (Button) view.findViewById(R.id.logout);
-        userPrefButton = (Button) view.findViewById(R.id.user_pref_button);
 
+        setMenuButtonOnClickListener();
+        setUpAddButton();
 
-
-        final Intent intent = new Intent(getActivity(), LoginPage.class);
-
-        userPrefButton.setOnClickListener(new View.OnClickListener(){
-            @Override
-            public void onClick(View view){
-                android.support.v4.app.Fragment UserPref = new UserPreferencesPage();
-
-                FragmentTransaction transaction = getFragmentManager().beginTransaction();
-                transaction.replace(R.id.my_movement_tab, UserPref);
-                transaction.commit();
-
-            }
-        });
-
-        logoutButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                LoginManager.getInstance().logOut();
-                startActivity(intent);
-            }
-        });
         setUpChangeFragmentsButton(view, new RecommendedMovementsTab(), R.id.movements);
 //        setUpChangeFragmentsButton(view, new AddMovementPage(), R.id.goto_add_movement_button);
-        ImageButton addMovementBtn = (ImageButton) view.findViewById(R.id.goto_add_movement_btn);
-        addMovementBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                ((MainActivity) getActivity()).changeFragment(new AddMovementPage());
-            }
-        });
+
+
         setUsersMovementsListener();
         if (!movements.isEmpty() && movements != null) {
             setUpListView();
