+++ Beakon/app/src/main/res/layout/fragment_add_movement_page.xml
@@ -4,39 +4,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:id="@+id/tab_buttons">
-
-        <Button
-            android:layout_width="0dp"
-            android:layout_weight="0.5"
-            android:layout_margin="0dp"
-            android:padding="0dp"
-            android:layout_height="wrap_content"
-            android:text="My Movements"
-            android:id="@+id/my_movements"
-            android:background="@color/colorAccentLight" />
-
-        <Button
-            android:layout_width="0dp"
-            android:layout_weight="0.5"
-            android:layout_margin="0dp"
-            android:padding="0dp"
-            android:layout_height="wrap_content"
-            android:text="Suggested"
-            android:id="@+id/this_page"
-            android:background="@color/colorBackground"/>
-
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="LOGOUT"/>
-    </LinearLayout>
-
     <ImageButton
         style="?android:attr/buttonStyleSmall"
         android:layout_width="@dimen/image_width"
