+++ Beakon/app/src/main/res/layout/fragment_my_movements_tab.xml
@@ -9,39 +9,10 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <LinearLayout
-            android:layout_width="match_parent"
+        <include layout="@layout/top_tabs"
+            android:id="@+id/tab_buttons"
             android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            android:id="@+id/tab_buttons">
-
-            <Button
-                android:layout_width="0dp"
-                android:layout_weight="0.5"
-                android:layout_margin="0dp"
-                android:padding="0dp"
-                android:layout_height="wrap_content"
-                android:text="My Movements"
-                android:id="@+id/my_movements"
-                android:background="@color/colorBackground" />
-
-            <Button
-                android:layout_width="0dp"
-                android:layout_weight="0.5"
-                android:layout_margin="0dp"
-                android:padding="0dp"
-                android:layout_height="wrap_content"
-                android:text="Suggested"
-                android:id="@+id/movements"
-                android:background="@color/colorAccentLight"/>
-
-            <Button
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="LOGOUT"
-                android:id="@+id/logout"/>
-
-        </LinearLayout>
+            android:layout_width="match_parent"/>
 
         <TextView
             android:id="@+id/no_movments_message"
@@ -49,7 +20,6 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="@string/no_my_movements"
-            android:layout_below="@+id/tab_buttons"
             android:textSize="30sp"
             android:padding="30dp"
             android:layout_marginTop="30dp"
@@ -60,25 +30,10 @@
             android:layout_width="match_parent"
             android:layout_height="456dp"
             android:id="@+id/my_movements_list"
-            android:layout_below="@id/tab_buttons"></ListView>
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Settings"
-            android:id="@+id/user_pref_button"
-            android:layout_alignBottom="@+id/my_movements_list"
-            android:layout_alignParentRight="true"
-            android:layout_alignParentEnd="true" />
+            android:layout_below="@+id/tab_buttons"></ListView>
 
     </RelativeLayout>
 
-    <ImageButton
-        android:layout_width="@dimen/image_width"
-        android:layout_height="@dimen/image_height"
-        android:id="@+id/goto_add_movement_btn"
-        android:src="@drawable/add"
-        android:background="@null"
-        android:layout_gravity="right|bottom" />
+    <include layout="@layout/bottom_buttons"/>
 
 </FrameLayout>
