+++ Beakon/app/src/main/res/layout/fragment_recommended_movements.xml
@@ -2,53 +2,26 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="erica.beakon.Pages.RecommendedMovementsTab">
+    tools:context="erica.beakon.Pages.RecommendedMovementsTab"
+    android:id="@+id/recommended_movements_tab">
+
 
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <LinearLayout
-            android:layout_width="match_parent"
+        <include layout="@layout/top_tabs"
+            android:id="@+id/tab_buttons"
             android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            android:id="@+id/tab_buttons">
-
-            <Button
-                android:layout_width="0dp"
-                android:layout_weight="0.5"
-                android:layout_margin="0dp"
-                android:padding="0dp"
-                android:layout_height="wrap_content"
-                android:text="My Movements"
-                android:id="@+id/my_movements"
-                android:background="@color/colorAccentLight" />
-
-            <Button
-                android:layout_width="0dp"
-                android:layout_weight="0.5"
-                android:layout_margin="0dp"
-                android:padding="0dp"
-                android:layout_height="wrap_content"
-                android:text="Suggested"
-                android:id="@+id/movements"
-                android:background="@color/colorBackground"/>
-
-
-            <Button
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="LOGOUT"
-                android:id="@+id/logout"/>
-        </LinearLayout>
+            android:layout_width="match_parent"/>
 
         <TextView
             android:id="@+id/no_rec_movments_message"
             android:visibility="invisible"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="@string/no_rec_movements"
             android:layout_below="@+id/tab_buttons"
+            android:text="@string/no_rec_movements"
             android:textSize="30sp"
             android:padding="30dp"
             android:layout_marginTop="30dp"
@@ -59,8 +32,8 @@
             android:layout_height="wrap_content"
             android:layout_margin="10dp"
             android:textSize="20sp"
-            android:id="@+id/nearby_movements_label"
             android:layout_below="@+id/tab_buttons"
+            android:id="@+id/nearby_movements_label"
             android:labelFor="@+id/recommended_nearby_movements_list"
             android:text="Nearby Movements"/>
 
@@ -90,5 +63,6 @@
 
     </RelativeLayout>
 
+    <include layout="@layout/bottom_buttons"/>
 
 </FrameLayout>
