+++ Beakon/app/src/main/res/layout/top_tabs.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent" android:layout_height="match_parent">
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        android:id="@+id/tab_buttons">
+
+        <Button
+            android:layout_width="0dp"
+            android:layout_weight="0.5"
+            android:layout_margin="0dp"
+            android:padding="0dp"
+            android:layout_height="wrap_content"
+            android:text="My Movements"
+            android:id="@+id/my_movements"
+            android:background="@color/colorBackground" />
+
+        <Button
+            android:layout_width="0dp"
+            android:layout_weight="0.5"
+            android:layout_margin="0dp"
+            android:padding="0dp"
+            android:layout_height="wrap_content"
+            android:text="Suggested"
+            android:id="@+id/movements"
+            android:background="@color/colorAccentLight"/>
+
+    </LinearLayout>
+</RelativeLayout>
