+++ Beakon/app/src/main/res/menu/popupmenu.xml
@@ -1,0 +1,11 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android" >
+
+    <item
+        android:id="@+id/one"
+        android:title="Settings"/>
+
+    <item
+        android:id="@+id/two"
+        android:title="Logout"/>
+
+</menu>
